[Summary]商品价格比较

+++ res/yaoyy/ms-service/src/main/java/com/ms/service/impl/CommodityServiceImpl.java
@@ -198,7 +198,7 @@ public class CommodityServiceImpl extends AbsCommonService<Commodity> implements
             throw new RuntimeException("商品不存在");
         }
         // 比较当前价格和之前价格是否有变动
-        if (!(vo.getPrice() == oldCommodity.getPrice())){
+        if (!(vo.getPrice().floatValue() == oldCommodity.getPrice().floatValue())){
             // 以前的价格存为历史价格
             HistoryPriceVo historyPrice=new HistoryPriceVo();
             historyPrice.setCommodityId(oldCommodity.getId());
